modify killAllCrabJobs script to take a directory as input

+++ multilep/test/scripts/killAllCrabJobs.sh
@@ -1,22 +1,19 @@
 #!/bin/bash
 echo "Killing all running crab jobs listed in ./crab directory"
-cd ../crab
-for dir in *
-    do if [[ -d $dir ]]; then
-        cd $dir
-        for dir2 in *
-            do if [[ -d ${dir2} ]]; then
-                cd ${dir2}
-                for dir3 in *
-                    do if [[ -d ${dir3} ]]; then
-                        cd ${dir3}
-                        crab kill --dir=./
-                        cd ..
-                    fi
-                done
-                cd ..
-            fi
-        done
-        cd ..
-    fi
-done
+if [[ -d $1 ]]; then
+    cd $1
+    for dir2 in *
+        do if [[ -d ${dir2} ]]; then
+            cd ${dir2}
+            for dir3 in *
+                do if [[ -d ${dir3} ]]; then
+                    cd ${dir3}
+                    crab kill --dir=./
+                    cd ..
+                fi
+            done
+            cd ..
+        fi
+    done
+    cd ..
+fi
